Move serializers to models.py
